Create and move some code into DistanceMatrixService
